Use smaller data types in cs_decode.
Old total .o size: 2354, 2450
New total .o size: 1596, 1678

+++ binary/custom-encoding/cs_decode/cs_decode.h
@@ -11,9 +11,19 @@ enum {
     BUILTIN_END
 };
 
+/*
+ * There are a bunch of uint8_t below, to save space in the binary. With larger
+ * schemas that will have to be uint16_t. (Maybe even uint32_t for some huge
+ * schemas.)
+ * The goal is to have tool.py generate this entire file, and generate the
+ * appropriate types based on the schema itself. Until that is done, I'm just
+ * changing the types to be the smallest. The compiler will warn/error when
+ * generated constants don't fit in these types.
+ */
+
 typedef struct {
-    unsigned code;
-    unsigned type;
+    uint8_t code;
+    uint8_t type;
 } cs_typedef_entry_t;
 
 #define CS_FLAG_REPEATABLE      1
@@ -21,12 +31,12 @@ typedef struct {
 typedef uint8_t cs_typedef_flags_t;
 
 typedef struct {
-    unsigned entry_count;
-    unsigned entry_index;
+    uint8_t entry_count;
+    uint8_t entry_index;
 } cs_typedef_t;
 
 typedef struct {
-    unsigned type_count;
+    uint8_t type_count;
     const cs_typedef_t *types;
     const cs_typedef_entry_t *all_entries;
     const cs_typedef_flags_t *all_flags;
@@ -34,7 +44,7 @@ typedef struct {
 
 typedef struct {
     unsigned values[10];
-    unsigned depth;
+    uint8_t depth;
 } cs_path_t;
 
 int cs_decode(
